+++ src/graph.cpp
@@ -15,9 +15,9 @@
 #include <iostream>
 
 int Graph::intersection_size(int v1,int v2) {
-    unsigned int l1, r1;
+    long long l1, r1;
     this->get_edge_index(v1, l1, r1);
-    unsigned int l2, r2;
+    long long l2, r2;
     this->get_edge_index(v2, l2, r2);
     int ans = 0;
     while(l1 < r1 && l2 < r2) {
@@ -43,7 +43,7 @@ int Graph::intersection_size_mpi(int v1, int v2) {
     int ans = 0;
     if (gm.include(v2))
         return intersection_size(v1, v2);
-    unsigned int l1, r1;
+    long long l1, r1;
     this->get_edge_index(v1, l1, r1);
     int *data = gm.getneighbor(v2);
     for (int l2 = 0; l1 < r1 && ~data[l2];) {
@@ -63,9 +63,9 @@ int Graph::intersection_size_mpi(int v1, int v2) {
 }
 
 int Graph::intersection_size_clique(int v1,int v2) {
-    unsigned int l1, r1;
+    long long l1, r1;
     this->get_edge_index(v1, l1, r1);
-    unsigned int l2, r2;
+    long long l2, r2;
     this->get_edge_index(v2, l2, r2);
     int min_vertex = v2;
     int ans = 0;
@@ -97,7 +97,7 @@ long long Graph::triangle_counting() {
     long long ans = 0;
     for(int v = 0; v < v_cnt; ++v) {
         // for v in G
-        unsigned int l, r;
+        long long l, r;
         this->get_edge_index(v, l, r);
         for(unsigned int v1 = l; v1 < r; ++v1) {
             //for v1 in N(v)
@@ -122,7 +122,7 @@ void Graph::tc_mt(long long *global_ans) {
     #pragma omp for schedule(dynamic)
     for(int v = 0; v < v_cnt; ++v) {
         // for v in G
-        unsigned int l, r;
+        long long l, r;
         this->get_edge_index(v, l, r);
         for(unsigned int v1 = l; v1 < r; ++v1) {
             if (v <= edge[v1])
@@ -171,7 +171,7 @@ long long Graph::triangle_counting_mpi(int thread_count) {
     return -1;
 }
 
-void Graph::get_edge_index(int v, unsigned int& l, unsigned int& r) const
+void Graph::get_edge_index(int v, long long & l, long long & r) const
 {
     // assert(!(is_local_graph));
     l = vertex[v];
@@ -211,12 +211,12 @@ void Graph::pattern_matching_func(const Schedule& schedule, VertexSet* vertex_se
         if (!clique)
             if (subtraction_set.has_data(vertex))
                 continue;
-        unsigned int l, r;
+        long long l, r;
         this->get_edge_index(vertex, l, r);
         bool is_zero = false;
         for (int prefix_id = schedule.get_last(depth); prefix_id != -1; prefix_id = schedule.get_next(prefix_id))
         {
-            vertex_set[prefix_id].build_vertex_set(schedule, vertex_set, &edge[l], (int)r - l, prefix_id, vertex, clique);
+            vertex_set[prefix_id].build_vertex_set(schedule, vertex_set, &edge[l], (int)(r - l), prefix_id, vertex, clique);
             if( vertex_set[prefix_id].get_size() == 0) {
                 is_zero = true;
                 break;
@@ -250,10 +250,10 @@ long long Graph::pattern_matching(const Schedule& schedule, int thread_count, bo
  #pragma omp for schedule(dynamic, 1) nowait
         for (int vertex = 0; vertex < v_cnt; ++vertex)
         {
-            unsigned int l, r;
+            long long l, r;
             this->get_edge_index(vertex, l, r);
             int prefix_id = schedule.get_last(0);
-            vertex_set[prefix_id].build_vertex_set(schedule, vertex_set, bs, &edge[l], (int)r - l, prefix_id, 0);
+            vertex_set[prefix_id].build_vertex_set(schedule, vertex_set, bs, &edge[l], (int)(r - l), prefix_id, 0);
             clique_matching_func(schedule, vertex_set, bs, local_ans, 1);
             int  *data = vertex_set[prefix_id].get_data_ptr();
             for(int i = 0; i < vertex_set[prefix_id].get_size(); i++) bs->dec(data[i]);
@@ -273,7 +273,7 @@ void Graph::clique_matching_func(const Schedule& schedule, VertexSet* vertex_set
 
     for (int i = 0; i < loop_size; ++i) {
         int vertex = loop_data_ptr[i];
-        unsigned int l, r;
+        long long l, r;
         this->get_edge_index(vertex, l, r);
         int prefix_id = schedule.get_last(depth);// only one prefix
         if(depth == schedule.get_size() - 2)
@@ -376,7 +376,7 @@ void Graph::pattern_matching_aggressive_func(const Schedule& schedule, VertexSet
         int vertex = loop_data_ptr[i];
         if (subtraction_set.has_data(vertex))
             continue;
-        unsigned int l, r;
+        long long l, r;
         this->get_edge_index(vertex, l, r);
         bool is_zero = false;
         for (int prefix_id = schedule.get_last(depth); prefix_id != -1; prefix_id = schedule.get_next(prefix_id))
@@ -431,7 +431,7 @@ long long Graph::pattern_matching_mpi(const Schedule& schedule, int thread_count
                 if ((range = gm.get_vertex_range()).first == -1) break;
                 //for (int vertex = v_cnt - range.second; vertex < v_cnt - range.first; vertex++) {//backwards slower than forwards
                 for (int vertex = range.first; vertex < range.second; vertex++) {
-                    unsigned int l, r;
+                    long long l, r;
                     this->get_edge_index(vertex, l, r);
                     match_start_vertex(vertex, edge + l, r - l);
                 }
@@ -487,7 +487,7 @@ void Graph::pattern_matching_aggressive_func_mpi(const Schedule& schedule, Verte
         Graphmpi &gm = Graphmpi::getinstance();
         //if (gm.include(vertex)) {
         if (true) {
-            unsigned int l, r;
+            long long l, r;
             this->get_edge_index(vertex, l, r);
             data = edge + l;
             size = r - l;
